Added Base XCode project file for MacOS.
This is the XCode project file for MacOS.
 
The file is a zip by default, and contains all the required data to launch a C++20 application automatically. 
XCode will automatically extract all the required data from this file.

Currently all tests are passing, and no code changes are required for the main repo

Please note, the zip file only contains data pretending to the project itself. It does not contain source code.

The project is reset to reference the current src and test folders in the main repo.

Requirements: Version 14.3.1 (14E300c) or higer

Older version may work, but the user will need to set this up themselves